requirements.txt
asgiref==3.5.2
astroid==2.11.6
attrs==22.1.0
awsebcli==3.20.3
bcrypt==4.0.0
beautifulsoup4==4.11.1
blessed==1.19.1
botocore==1.23.54
cached-property==1.5.2
cement==2.8.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.3
coverage==6.4.4
crispy-bootstrap5==0.6
cryptography==37.0.4
defusedxml==0.7.1
dill==0.3.5.1
dj-database-url==1.0.0
Django==4.0.6
django-allauth==0.51.0
django-autoslug==1.9.8
django-bootstrap-v5==1.0.11
django-cleanup==6.0.0
django-countries==7.3.2
django-crispy-forms==1.14.0
django-filter==22.1
django-heroku==0.3.1
django-localflavor==3.1
django-multiselectfield==0.1.12
django-phone-field==1.8.1
django-pwa==1.0.10
django-storages==1.13.1
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
future==0.16.0
gunicorn==20.1.0
idna==3.3
isort==5.10.1
jmespath==0.10.0
jsonschema==3.2.0
lazy-object-proxy==1.7.1
mccabe==0.7.0
oauthlib==3.2.0
paramiko==2.11.0
pathspec==0.9.0
platformdirs==2.5.2
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
pylint==2.14.4
pylint-django==2.5.3
pylint-plugin-utils==0.7
PyNaCl==1.5.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python-stdnum==1.17
python3-openid==3.2.0
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.1
semantic-version==2.8.5
six==1.14.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
termcolor==1.1.0
texttable==1.6.4
tomli==2.0.1
tomlkit==0.11.1
typing_extensions==4.3.0
urllib3==1.26.11
wcwidth==0.1.9
websocket-client==0.59.0
whitenoise==6.2.0
wrapt==1.14.1
